requirements.txt
Pillow>=9.0.0
numpy>=1.22.0
opencv-python>=4.5.0
scipy>=1.8.0
requests>=2.27.0